scoutlens/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0
pydantic-settings>=2.0
email-validator>=2.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.28.0
psycopg2-binary>=2.9
alembic>=1.12
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
pandas>=2.1
pyarrow>=14.0
numpy>=1.25
statsbombpy>=1.11
scikit-learn>=1.3
faiss-cpu>=1.7
umap-learn>=0.5
shap>=0.42
pytest>=7.4
pytest-asyncio>=0.21
httpx>=0.24
ruff>=0.1
black>=23.0
